Implement `VacantEntry` and `OccupiedEntry`
TODO: Implement `VacantEntry::insert`.

The problem with `VacantEntry::insert` is that `SgMap::insert` returns
`Option<V>` i.e. `Some` if old value returned, else `None`; this means
that we need to look for the new node again.
This is problematic since the key has been moved into the map and cannot
be used again.